Moved sequence display components to its own package
