+++ modules/edu.gemini.seqexec.web/edu.gemini.seqexec.web.client/src/main/scala/edu/gemini/seqexec/web/client/components/sequence/SequenceArea.scala
@@ -1,50 +1,14 @@
-package edu.gemini.seqexec.web.client.components
+package edu.gemini.seqexec.web.client.components.sequence
 
+import edu.gemini.seqexec.web.client.components.{TabularMenu, TextMenuSegment}
 import edu.gemini.seqexec.web.client.components.TabularMenu.TabItem
 import edu.gemini.seqexec.web.client.model.{SeqexecCircuit, SequenceTab, SequencesOnDisplay}
 import edu.gemini.seqexec.web.client.semanticui._
 import edu.gemini.seqexec.web.client.semanticui.elements.button.Button
-import japgolly.scalajs.react.{ReactComponentB, ReactElement}
-import japgolly.scalajs.react.vdom.prefix_<^._
 import edu.gemini.seqexec.web.client.semanticui.elements.icon.Icon.{IconCaretRight, IconInbox}
 import edu.gemini.seqexec.web.client.semanticui.elements.message.IconMessage
-
-object HeadersSideBar {
-  val component = ReactComponentB[Unit]("HeadersSideBar")
-    .stateless
-    .render(_ =>
-      <.div(
-        ^.cls := "ui raised secondary segment",
-        <.h4("Headers"),
-        <.div(
-          ^.cls := "ui form",
-          <.div(
-            ^.cls := "required field",
-            <.label("Observer"),
-            <.input(
-              ^.`type` :="text",
-              ^.autoComplete :="off"
-            )
-          ),
-          <.div(
-            ^.cls := "required field",
-            <.label("SOS"),
-            <.input(
-              ^.`type` :="text",
-              ^.autoComplete :="off"
-            )
-          ),
-          DropdownMenu(DropdownMenu.Props("Image Quality", "Select", List("IQ20", "IQ70", "IQ85", "Any"))),
-          DropdownMenu(DropdownMenu.Props("Cloud Cover", "Select", List("CC20", "CC50", "CC70", "CC80", "CC90", "Any"))),
-          DropdownMenu(DropdownMenu.Props("Water Vapor", "Select", List("WV20", "WV50", "WV80", "Any"))),
-          DropdownMenu(DropdownMenu.Props("Sky Background", "Select", List("SB20", "SB50", "SB80", "Any")))
-        )
-      )
-    )
-    .build
-
-  def apply() = component()
-}
+import japgolly.scalajs.react.vdom.prefix_<^._
+import japgolly.scalajs.react.{ReactComponentB, ReactElement}
 
 object SequenceContainer {
   val component = ReactComponentB[Unit]("HeadersSideBar")
@@ -81,27 +45,6 @@ object SequenceContainer {
                 )
               )
             )
-          )
-        )
-      )
-    )
-    .build
-
-  def apply() = component()
-}
-
-object LogArea {
-  val component = ReactComponentB[Unit]("LogArea")
-    .stateless
-    .render(_ =>
-      <.div(
-        ^.cls := "ui raised secondary segment",
-        <.h4("Log"),
-        <.div(
-          ^.cls := "ui form",
-          <.div(
-            ^.cls := "field",
-            <.textarea()
           )
         )
       )
